Moving files to library
Updating library classes
