+++ S4GFXLibrary/GFX/IGfxImage.cs
@@ -1,6 +1,12 @@
-﻿namespace S4GFXFramework.GFX
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace S4GFXLibrary.GFX
 {
-    public interface IGfxImage
+    interface IGfxImage
     {
         ImageData GetImageData();
         int GetDataSize();
@@ -12,5 +18,7 @@
 
         int Flag1 { get; set; }
         int Flag2 { get; set; }
+
+        byte[] GetData();
     }
 }
